requirements.txt
pymongo
plotly
dash
pymongo
pandas
dash-bootstrap-components
python-dotenv